app/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
numpy
opencv-python
ultralytics
torch
pillow
faiss-cpu
